Factor out EndpointsNamer logging code for reuse
